fix: remove professores files and adding routes to app file
