--- database/professores.py
@@ -1,8 +1,0 @@
-professores = [
-    {"id": 1, "nome": "Ana Silva", "disciplina": "Matemática"},
-    {"id": 2, "nome": "Carlos Souza", "disciplina": "História"},
-    {"id": 3, "nome": "Fernanda Costa", "disciplina": "Biologia"}
-]
-
-
-
