--- routes/professores_routes.py
@@ -1,78 +1,0 @@
-from flask import Blueprint, jsonify, request
-
-from database.professores import professores
-
-professores_blueprint = Blueprint('professores', __name__)
-
-
-@professores_blueprint.route('/professores', methods=['GET'])
-def get_professores():
-    data = professores
-    if data is None:
-        return jsonify([])
-    return jsonify(professores)
-
-
-@professores_blueprint.route('/professor/<int:professor_id>', methods=['GET'])
-def get_professor(professor_id):
-    for professor in professores:
-        if professor['id'] == professor_id:
-            return jsonify({
-                'professor': professor
-            })
-
-    return jsonify({'error': 'professor nao encontrado'}), 404
-
-
-@professores_blueprint.route('/professor', methods=['POST'])
-def create_professor():
-    data = request.json
-    if 'nome' not in data:
-        return jsonify({'error': 'nome é um campo obrigatório'}), 400
-
-    if 'disciplina' not in data:
-        return jsonify({'error': 'disciplina é um campo obrigatório'}), 400
-
-    professor = {
-        'id': len(professores) + 1,
-        'nome': data['nome'],
-        'disciplina': data['disciplina']
-    }
-    professores.append(professor)
-    return jsonify({
-        'message': 'professor criado com sucesso',
-        'professor': professor
-    })
-
-
-@professores_blueprint.route('/professor/<int:professor_id>', methods=['PUT'])
-def update_professor(professor_id):
-    for professor in professores:
-        if professor['id'] == professor_id:
-            data = request.json
-            if 'nome' not in data:
-                return jsonify({'error': 'nome é um campo obrigatório'}), 400
-
-            if 'disciplina' not in data:
-                return jsonify({'error': 'disciplina é um campo obrigatório'}), 400
-
-            professor["nome"] = data['nome']
-            professor["disciplina"] = data['disciplina']
-            return jsonify({
-                'message': 'professor atualizado com sucesso',
-                'professor': professor
-            })
-
-    return jsonify({'error': 'professor nao encontrado'}), 404
-
-
-@professores_blueprint.route('/professor/<int:professor_id>', methods=['DELETE'])
-def delete_professor(professor_id):
-    for professor in professores:
-        if professor['id'] == professor_id:
-            professores.remove(professor)
-            return jsonify({
-                'message': 'professor removido com sucesso'
-            })
-
-    return jsonify({'error': 'professor nao encontrado'}), 404
